research/requirements-research.txt
notebook
faiss-cpu
sentence-transformers
pinecone-client
pandas
tqdm
Pillow
requests
